Added service_handlers_registrars package TeacherInfoServiceHandlersRegistrar.py renamed into teacher_info_service_handlers_registrar.py and placed in service_handlers_registrars
